Move event store badger impl in with that server interface.
This is a weird package hierarchy, but we're going to delete that whole
package soon & this just makes it easier to find all the relevant
usages.
