+++ internal/logd/server/store/badger/encoding.go
