+++ internal/logd/server/store/badger/history.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/deref/exo/internal/logd/api"
-	"github.com/deref/exo/internal/logd/store"
+	"github.com/deref/exo/internal/logd/server/store"
 	"github.com/deref/exo/internal/util/binaryutil"
 	"github.com/dgraph-io/badger/v3"
 )
