+++ internal/logd/server/store/badger/log.go
