+++ internal/logd/server/store/badger/sink.go
