+++ internal/logd/server/store/badger/store.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/deref/exo/internal/gensym"
-	"github.com/deref/exo/internal/logd/store"
+	"github.com/deref/exo/internal/logd/server/store"
 	"github.com/deref/exo/internal/util/logging"
 	"github.com/dgraph-io/badger/v3"
 )
