+++ internal/logd/server/store/cursor.go
