+++ internal/logd/server/store/cursor_test.go
@@ -3,7 +3,7 @@ package store_test
 import (
 	"testing"
 
-	"github.com/deref/exo/internal/logd/store"
+	"github.com/deref/exo/internal/logd/server/store"
 	"github.com/stretchr/testify/assert"
 )
